Moves shortcodes and filters to _client/config
