+++ _client/config/shortcode-figure.js
